別ファイルへの切り出し
